refactor(calendar): move CalendarUtil to domain layer
